+++ app/src/models/nic_model.py
