+++ app/src/models/student_loan_model.py
@@ -1,6 +1,5 @@
 # Student Loan Repayment Class
 class StudentLoanModel():    
-    
     # payment free amounts
     payment_plan1_threshold = 18330
     payment_plan2_threshold = 25000
