+++ app/views/base.html
@@ -31,7 +31,7 @@
 			<div id=footer>
 				<section class="container">
 					<p>
-						Developed by: <span><a href="https://github.com/jacklovett">Jack Lovett</a></span>
+						Developed by: <a href="https://github.com/jacklovett">Jack Lovett</a> - 2018
 					</p> 
 				</section>
 			</div>
